// src/core/database/mongo/schema/product-category.schema.ts
import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
import { HydratedDocument } from 'mongoose';
import { AutoGenerated } from 'src/core/decorators/auto-generated.decorator';
import {
  ProductCategoryStatus,
  ProductCategoryTypes,
} from 'src/shared/enums/product-category.enums';

export type ProductCategoryDocument = HydratedDocument<ProductCategory>;

@Schema({
  collection: 'product_categories',
})
export class ProductCategory {
  /* ======================================================
   * IDENTITY
   * ====================================================== */

  @AutoGenerated()
  @Prop({ required: true, unique: true, type: String, trim: true })
  categoryId!: string;

  /* ======================================================
   * HIERARCHY
   * ====================================================== */

  @Prop({
    type: String,
    trim: true,
    default: null,
    validate: {
      validator: function (value: string) {
        // ✅ parentId required only for CHILD
        if (this.type === ProductCategoryTypes.CHILD) {
          return !!value;
        }
        return true;
      },
      message: 'parentId is required when type is CHILD',
    },
  })
  parentId!: string;

  @Prop({
    required: true,
    enum: ProductCategoryTypes,
    type: String,
  })
  type!: string;

  /* ======================================================
   * NAME
   * ====================================================== */

  @Prop({
    required: true,
    type: String,
    trim: true,
    minlength: 2,
    maxlength: 100,
  })
  name!: string;

  /* ======================================================
   * STATUS
   * ====================================================== */

  @Prop({
    type: String,
    enum: ProductCategoryStatus,
    default: ProductCategoryStatus.ACTIVE,
  })
  status!: ProductCategoryStatus;
}

export const ProductCategorySchema = SchemaFactory.createForClass(ProductCategory);